Interest computer
Design a small program to input the principal and interest rate. The default principal is 10000 yuan and the interest rate is 3%. Calculate how much you get after five years when you deposit it in the bank and calculate the compound interest? (add the principal and interest as the new principal every 1 year) (realized by using the for loop)